ObjectType and old Animation logic removed for a more encaspsulated logic (it still needs to be refactored even more as this approach is still not ideal). Mario walking animation seems to be the only problem. Further development should achieve frame-independent, time-dependent animations and fix the existing mario walk animation issue.
